Subset wells to region

Filter the GWELLS data returning only wells within the provided shapefile.

Usage

wells_subset(region, fix_bottom = TRUE, fix_depth = TRUE, update = FALSE)

Arguments

region

sf simple features object. Shape file of the region of interest.

fix_bottom

Logical. Whether to add 1m to bottom lithology intervals that has no thickness (identified by flat_int_bottom). Default TRUE.

fix_depth

Logical. Whether to fix missing well depths by making them equal to the depth of the final lithology layer. Default TRUE.

update

Logical. Force update of the data?

Examples

if (FALSE) { # interactive()

library(sf)

# Load a shape file defining the region of interest
creek_sf <- st_read("misc/data/Clinton_Creek.shp")

# Get wells within this region
creek_wells <- wells_subset(creek_sf)
}